moved and removed resource bundle
